+++ metaserver/app/config.go
@@ -25,12 +25,13 @@ type SectionPID struct {
 type SectionCore struct {
 	Mode            string     `yaml:"mode"`
 	FailFastTimeout int        `yaml:"fail_fast_timeout"`
+	BindAddr        string     `yaml:"bind_addr"`
 	PID             SectionPID `yaml:"pid"`
 }
 
-// SectionKafka is sub section of config.
+// SectionRedis is sub section of config.
 type SectionRedis struct {
-	Sentinels        []string `yaml:"sentinel"`
+	Sentinels        []string `yaml:"sentinels"`
 	MetaDBName       string   `yaml:"meta_db_name"`
 	UpdateInterval   int      `yaml:"update_interval"`
 	MetaHashtable    string   `yaml:"meta_hashtable"`
@@ -43,13 +44,11 @@ func LoadConfYaml(confPath string) (ConfYaml, error) {
 	var config ConfYaml
 
 	configFile, err := ioutil.ReadFile(confPath)
-
 	if err != nil {
 		return config, err
 	}
 
 	err = yaml.Unmarshal(configFile, &config)
-
 	if err != nil {
 		return config, err
 	}