+++ metaserver/change_log.md
@@ -1,6 +1,15 @@
 ## develop history ##
 ---
 
+-2017/09/12
+   > feature
+   * use gogoprotobuf 
+
+-2017/08/27
+   > feature
+   * use transaction when store meta data
+   * load meta data from metaDB when app startup
+
 - 2017/08/14
    > feature
 	* store meta data to metaDB